+++ modules/AWSPowerShell/Cmdlets/AmazonAPIGateway/Basic/New-AGStage-Cmdlet.cs
@@ -87,6 +87,16 @@ namespace Amazon.PowerShell.Cmdlets.AG
         public System.String StageName { get; set; }
         
         /// <summary>
+        /// <para>
+        /// <para>A map that defines the stage variables for the new <a>Stage</a> resource. Variable
+        /// names can have alphabetic characters, and the values must match [A-Za-z0-9-._~:/?#&amp;=,]+</para>
+        /// </para>
+        /// </summary>
+        [System.Management.Automation.Parameter]
+        [Alias("Variables")]
+        public System.Collections.Hashtable Variable { get; set; }
+        
+        /// <summary>
         /// This parameter overrides confirmation prompts to force 
         /// the cmdlet to continue its operation. This parameter should always
         /// be used with caution.
@@ -118,6 +128,14 @@ namespace Amazon.PowerShell.Cmdlets.AG
             context.Description = this.Description;
             context.RestApiId = this.RestApiId;
             context.StageName = this.StageName;
+            if (this.Variable != null)
+            {
+                context.Variables = new Dictionary<System.String, System.String>(StringComparer.Ordinal);
+                foreach (var hashKey in this.Variable.Keys)
+                {
+                    context.Variables.Add((String)hashKey, (String)(this.Variable[hashKey]));
+                }
+            }
             
             var output = Execute(context) as CmdletOutput;
             ProcessOutput(output);
@@ -154,6 +172,10 @@ namespace Amazon.PowerShell.Cmdlets.AG
             if (cmdletContext.StageName != null)
             {
                 request.StageName = cmdletContext.StageName;
+            }
+            if (cmdletContext.Variables != null)
+            {
+                request.Variables = cmdletContext.Variables;
             }
             
             CmdletOutput output;
@@ -196,6 +218,7 @@ namespace Amazon.PowerShell.Cmdlets.AG
             public System.String Description { get; set; }
             public System.String RestApiId { get; set; }
             public System.String StageName { get; set; }
+            public Dictionary<System.String, System.String> Variables { get; set; }
         }
         
     }