+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### 3.1.27.0 (2015-11-10)
+  * Amazon API Gateway
+    - Added suppport for stage variables to the New-AGDeployment, New-AGStage and Test-AGInvokeMethod cmdlets with a new parameter, -Variable, (of Hashtable type). Stage variables act like environment variables for use in your API setup and mapping templates and enable configuration of different deployment stages (e.g., alpha, beta, production) of your API.
+
 ### 3.1.26.0 (2015-11-07)
   * AWS IoT
     - Fixed an issue affecting New-IOTKeysAndCertificate, Update-IOTCertificate and others where 'signature mismatch' errors were being returned.